+++ README.md
@@ -1,29 +1,23 @@
-# shadowsocks-plus
+# shadowsocks-plus v2
 -----
-发布页面: https://www.ifxor.com/archives/297/
-基于 Shadowsocks Go 版本，加入了以下特性：
 
-- 启动后降低权限至 nobody , 增强安全性。
-- 与 kcptun 集成，配合相应客户端可加速传输。
-- 网页控制面板
+客户端和服务端使用方法与原版 shadowsocks-go 完全相同。
 
-支持系统： Debian/Ubuntu (32/64位)，CentOS (32位，64位需自行安装32位C和C++支持库)
+-----
 
-一键安装：
-rm ssplus_install.sh ; wget --no-check-certificate https://www.ifxor.com/installers/ssplus/ssplus_install.sh && bash ssplus_install.sh
+### 功能
+已实现：
+原版 shadowsocks-go 所有功能
+集成 kcptun 实现加速
+多端口支持
 
-运行：ssplus-server -c /etc/ssplus.conf
+Todo:
+支持动态增加/删除用户（端口）
+Android 客户端
 
-开机自启动：将以上命令加入 /etc/rc.local 即可。
+-----
 
-默认端口：6719
+### Copyright
+本项目采用 LGPL License.
+引用了 shadowsocks-go, kcptun 的代码。感谢相关项目的作者 :-)
 
-默认密码：SSplus_Default_Password
-
-加密算法：chacha20
-
-kcptun 加速端口为SS端口号 + 10000，即默认为 16719
-
-SS端口号范围 1000 – 9999。
-
-配置文件：/etc/ssplus.conf（格式和原版一样）
