--- src/shadowsocks-server/server.go
@@ -1,388 +1,0 @@
-package main
-
-import (
-	"bytes"
-	"encoding/binary"
-	"errors"
-	"flag"
-	"fmt"
-	"io"
-	"log"
-	"net"
-	"os"
-	"os/signal"
-	"runtime"
-	"strconv"
-	"strings"
-	"sync"
-	"syscall"
-
-	ss "github.com/shadowsocks/shadowsocks-go/shadowsocks"
-)
-
-// #cgo LDFLAGS: -lssplus
-// #include "ssplus.h"
-import "C"
-
-const (
-	idType  = 0 // address type index
-	idIP0   = 1 // ip addres start index
-	idDmLen = 1 // domain address length index
-	idDm0   = 2 // domain address start index
-
-	typeIPv4 = 1 // type is ipv4 address
-	typeDm   = 3 // type is domain address
-	typeIPv6 = 4 // type is ipv6 address
-
-	lenIPv4     = net.IPv4len + 2 // ipv4 + 2port
-	lenIPv6     = net.IPv6len + 2 // ipv6 + 2port
-	lenDmBase   = 2               // 1addrLen + 2port, plus addrLen
-	lenHmacSha1 = 10
-)
-
-var debug ss.DebugLog
-
-func getRequest(conn *ss.Conn, auth bool) (host string, ota bool, err error) {
-	ss.SetReadTimeout(conn)
-
-	// buf size should at least have the same size with the largest possible
-	// request size (when addrType is 3, domain name has at most 256 bytes)
-	// 1(addrType) + 1(lenByte) + 256(max length address) + 2(port) + 10(hmac-sha1)
-	buf := make([]byte, 270)
-	// read till we get possible domain length field
-	if _, err = io.ReadFull(conn, buf[:idType+1]); err != nil {
-		return
-	}
-
-	var reqStart, reqEnd int
-	addrType := buf[idType]
-	switch addrType & ss.AddrMask {
-	case typeIPv4:
-		reqStart, reqEnd = idIP0, idIP0+lenIPv4
-	case typeIPv6:
-		reqStart, reqEnd = idIP0, idIP0+lenIPv6
-	case typeDm:
-		if _, err = io.ReadFull(conn, buf[idType+1:idDmLen+1]); err != nil {
-			return
-		}
-		reqStart, reqEnd = idDm0, int(idDm0+buf[idDmLen]+lenDmBase)
-	default:
-		err = fmt.Errorf("addr type %d not supported", addrType&ss.AddrMask)
-		return
-	}
-
-	if _, err = io.ReadFull(conn, buf[reqStart:reqEnd]); err != nil {
-		return
-	}
-
-	// Return string for typeIP is not most efficient, but browsers (Chrome,
-	// Safari, Firefox) all seems using typeDm exclusively. So this is not a
-	// big problem.
-	switch addrType & ss.AddrMask {
-	case typeIPv4:
-		host = net.IP(buf[idIP0 : idIP0+net.IPv4len]).String()
-	case typeIPv6:
-		host = net.IP(buf[idIP0 : idIP0+net.IPv6len]).String()
-	case typeDm:
-		host = string(buf[idDm0 : idDm0+buf[idDmLen]])
-	}
-	// parse port
-	port := binary.BigEndian.Uint16(buf[reqEnd-2 : reqEnd])
-	host = net.JoinHostPort(host, strconv.Itoa(int(port)))
-	// if specified one time auth enabled, we should verify this
-	if auth || addrType&ss.OneTimeAuthMask > 0 {
-		ota = true
-		if _, err = io.ReadFull(conn, buf[reqEnd:reqEnd+lenHmacSha1]); err != nil {
-			return
-		}
-		iv := conn.GetIv()
-		key := conn.GetKey()
-		actualHmacSha1Buf := ss.HmacSha1(append(iv, key...), buf[:reqEnd])
-		if !bytes.Equal(buf[reqEnd:reqEnd+lenHmacSha1], actualHmacSha1Buf) {
-			err = fmt.Errorf("verify one time auth failed, iv=%v key=%v data=%v", iv, key, buf[:reqEnd])
-			return
-		}
-	}
-	return
-}
-
-const logCntDelta = 100
-
-var connCnt int
-var nextLogConnCnt int = logCntDelta
-
-func handleConnection(conn *ss.Conn, auth bool) {
-	var host string
-
-	connCnt++ // this maybe not accurate, but should be enough
-	if connCnt-nextLogConnCnt >= 0 {
-		// XXX There's no xadd in the atomic package, so it's difficult to log
-		// the message only once with low cost. Also note nextLogConnCnt maybe
-		// added twice for current peak connection number level.
-		log.Printf("Number of client connections reaches %d\n", nextLogConnCnt)
-		nextLogConnCnt += logCntDelta
-	}
-
-	// function arguments are always evaluated, so surround debug statement
-	// with if statement
-	if debug {
-		debug.Printf("new client %s->%s\n", conn.RemoteAddr().String(), conn.LocalAddr())
-	}
-	closed := false
-	defer func() {
-		if debug {
-			debug.Printf("closed pipe %s<->%s\n", conn.RemoteAddr(), host)
-		}
-		connCnt--
-		if !closed {
-			conn.Close()
-		}
-	}()
-
-	host, ota, err := getRequest(conn, auth)
-	if err != nil {
-		log.Println("error getting request", conn.RemoteAddr(), conn.LocalAddr(), err)
-		return
-	}
-	debug.Println("connecting", host)
-	remote, err := net.Dial("tcp", host)
-	if err != nil {
-		if ne, ok := err.(*net.OpError); ok && (ne.Err == syscall.EMFILE || ne.Err == syscall.ENFILE) {
-			// log too many open file error
-			// EMFILE is process reaches open file limits, ENFILE is system limit
-			log.Println("dial error:", err)
-		} else {
-			log.Println("error connecting to:", host, err)
-		}
-		return
-	}
-	defer func() {
-		if !closed {
-			remote.Close()
-		}
-	}()
-	if debug {
-		debug.Printf("piping %s<->%s ota=%v connOta=%v", conn.RemoteAddr(), host, ota, conn.IsOta())
-	}
-	if ota {
-		go ss.PipeThenCloseOta(conn, remote)
-	} else {
-		go ss.PipeThenClose(conn, remote)
-	}
-	ss.PipeThenClose(remote, conn)
-	closed = true
-	return
-}
-
-type PortListener struct {
-	password string
-	listener net.Listener
-}
-
-type PasswdManager struct {
-	sync.Mutex
-	portListener map[string]*PortListener
-}
-
-func (pm *PasswdManager) add(port, password string, listener net.Listener) {
-	pm.Lock()
-	pm.portListener[port] = &PortListener{password, listener}
-	pm.Unlock()
-}
-
-func (pm *PasswdManager) get(port string) (pl *PortListener, ok bool) {
-	pm.Lock()
-	pl, ok = pm.portListener[port]
-	pm.Unlock()
-	return
-}
-
-func (pm *PasswdManager) del(port string) {
-	pl, ok := pm.get(port)
-	if !ok {
-		return
-	}
-	pl.listener.Close()
-	pm.Lock()
-	delete(pm.portListener, port)
-	pm.Unlock()
-}
-
-// Update port password would first close a port and restart listening on that
-// port. A different approach would be directly change the password used by
-// that port, but that requires **sharing** password between the port listener
-// and password manager.
-func (pm *PasswdManager) updatePortPasswd(port, password string, auth bool) {
-	pl, ok := pm.get(port)
-	if !ok {
-		log.Printf("new port %s added\n", port)
-	} else {
-		if pl.password == password {
-			return
-		}
-		log.Printf("closing port %s to update password\n", port)
-		pl.listener.Close()
-	}
-	// run will add the new port listener to passwdManager.
-	// So there maybe concurrent access to passwdManager and we need lock to protect it.
-	go run(port, password, auth)
-}
-
-var passwdManager = PasswdManager{portListener: map[string]*PortListener{}}
-
-func updatePasswd() {
-	log.Println("updating password")
-	newconfig, err := ss.ParseConfig(configFile)
-	if err != nil {
-		log.Printf("error parsing config file %s to update password: %v\n", configFile, err)
-		return
-	}
-	oldconfig := config
-	config = newconfig
-
-	if err = unifyPortPassword(config); err != nil {
-		return
-	}
-	for port, passwd := range config.PortPassword {
-		passwdManager.updatePortPasswd(port, passwd, config.Auth)
-		if oldconfig.PortPassword != nil {
-			delete(oldconfig.PortPassword, port)
-		}
-	}
-	// port password still left in the old config should be closed
-	for port, _ := range oldconfig.PortPassword {
-		log.Printf("closing port %s as it's deleted\n", port)
-		passwdManager.del(port)
-	}
-	log.Println("password updated")
-}
-
-func waitSignal() {
-	var sigChan = make(chan os.Signal, 1)
-	signal.Notify(sigChan, syscall.SIGHUP)
-	for sig := range sigChan {
-		if sig == syscall.SIGHUP {
-			updatePasswd()
-		} else {
-			// is this going to happen?
-			log.Printf("caught signal %v, exit", sig)
-			os.Exit(0)
-		}
-	}
-}
-
-func run(port, password string, auth bool) {
-	ln, err := net.Listen("tcp", ":"+port)
-	if err != nil {
-		log.Printf("error listening port %v: %v\n", port, err)
-		os.Exit(1)
-	}
-	passwdManager.add(port, password, ln)
-	var cipher *ss.Cipher
-	log.Printf("server listening port %v ...\n", port)
-	for {
-		conn, err := ln.Accept()
-		if err != nil {
-			// listener maybe closed to update password
-			debug.Printf("accept error: %v\n", err)
-			return
-		}
-		// Creating cipher upon first connection.
-		if cipher == nil {
-			log.Println("creating cipher for port:", port)
-			cipher, err = ss.NewCipher(config.Method, password)
-			if err != nil {
-				log.Printf("Error generating cipher for port: %s %v\n", port, err)
-				conn.Close()
-				continue
-			}
-		}
-		go handleConnection(ss.NewConn(conn, cipher.Copy()), auth)
-	}
-}
-
-func enoughOptions(config *ss.Config) bool {
-	return config.ServerPort != 0 && config.Password != ""
-}
-
-func unifyPortPassword(config *ss.Config) (err error) {
-	if len(config.PortPassword) == 0 { // this handles both nil PortPassword and empty one
-		if !enoughOptions(config) {
-			fmt.Fprintln(os.Stderr, "must specify both port and password")
-			return errors.New("not enough options")
-		}
-		port := strconv.Itoa(config.ServerPort)
-		config.PortPassword = map[string]string{port: config.Password}
-	} else {
-		if config.Password != "" || config.ServerPort != 0 {
-			fmt.Fprintln(os.Stderr, "given port_password, ignore server_port and password option")
-		}
-	}
-	return
-}
-
-var configFile string
-var config *ss.Config
-
-func main() {
-	log.SetOutput(os.Stdout)
-
-	var cmdConfig ss.Config
-	var printVer bool
-	var core int
-
-	C.ssp_start()
-
-	flag.BoolVar(&printVer, "version", false, "print version")
-	flag.StringVar(&configFile, "c", "config.json", "specify config file")
-	flag.StringVar(&cmdConfig.Password, "k", "", "password")
-	flag.IntVar(&cmdConfig.ServerPort, "p", 0, "server port")
-	flag.IntVar(&cmdConfig.Timeout, "t", 300, "timeout in seconds")
-	flag.StringVar(&cmdConfig.Method, "m", "", "encryption method, default: aes-256-cfb")
-	flag.IntVar(&core, "core", 0, "maximum number of CPU cores to use, default is determinied by Go runtime")
-	flag.BoolVar((*bool)(&debug), "d", false, "print debug message")
-
-	flag.Parse()
-
-	if printVer {
-		ss.PrintVersion()
-		os.Exit(0)
-	}
-
-	ss.SetDebug(debug)
-
-	if strings.HasSuffix(cmdConfig.Method, "-auth") {
-		cmdConfig.Method = cmdConfig.Method[:len(cmdConfig.Method)-5]
-		cmdConfig.Auth = true
-	}
-
-	var err error
-	config, err = ss.ParseConfig(configFile)
-	if err != nil {
-		if !os.IsNotExist(err) {
-			fmt.Fprintf(os.Stderr, "error reading %s: %v\n", configFile, err)
-			os.Exit(1)
-		}
-		config = &cmdConfig
-	} else {
-		ss.UpdateConfig(config, &cmdConfig)
-	}
-	if config.Method == "" {
-		config.Method = "aes-256-cfb"
-	}
-	if err = ss.CheckCipherMethod(config.Method); err != nil {
-		fmt.Fprintln(os.Stderr, err)
-		os.Exit(1)
-	}
-	if err = unifyPortPassword(config); err != nil {
-		os.Exit(1)
-	}
-	if core > 0 {
-		runtime.GOMAXPROCS(core)
-	}
-	for port, password := range config.PortPassword {
-		go run(port, password, config.Auth)
-	}
-
-	waitSignal()
-}
