--- src/shadowsocks-server/ssplus.h
@@ -1,1 +1,0 @@
-extern void ssp_start();
